pyproject.toml
[project]
name = "memorious-mcp"
version = "0.1.0"
description = "Persistent semantic key-value memory MCP server with ChromaDB backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.21",
    "mcp[cli]>=1.14.0",
    "fastmcp>=2.0.0",
    "pytest>=8.4.2",
]
license = { text = "MIT" }
authors = [
  { name = "Cedric Vidal", email = "cedric.vidal@gmail.com" },
]

[project.scripts]
memorious-mcp = "memorious_mcp.main:main"

[build-system]
requires = ["uv_build >= 0.7.19, <0.9.0"]
build-backend = "uv_build"

[tool.uv]
package = true

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pyyaml>=6.0",
]

[project.urls]
Documentation = "https://github.com/cedricvidal/memorious-mcp#readme"
Issues = "https://github.com/cedricvidal/memorious-mcp/issues"
Source = "https://github.com/cedricvidal/memorious-mcp"
